Added batch size option

+++ Axodox.Common/Storage/UwpStorage.cpp
@@ -10,7 +10,7 @@ namespace Axodox::Storage
   std::vector<uint8_t> read_file(const winrt::Windows::Storage::StorageFile& file)
   {
     auto stream = file.OpenReadAsync().get();
-    auto buffer = stream.ReadAsync(Buffer{ uint32_t(stream.Size()) }, stream.Size(), InputStreamOptions::None).get();
+    auto buffer = stream.ReadAsync(Buffer{ uint32_t(stream.Size()) }, uint32_t(stream.Size()), InputStreamOptions::None).get();
 
     vector<uint8_t> result;
     result.resize(buffer.Length());
